Move redis init & move checks.py to utils/
